Simplify plug-ins further
