+++ csharp/test/Ice/plugin/Client.cs
@@ -18,17 +18,18 @@ namespace ZeroC.Ice.Test.Plugin
         {
             string pluginPath =
                 string.Format("msbuild/plugin/{0}/Plugin.dll",
-                            Path.GetFileName(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)));
+                              Path.GetFileName(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location)));
             {
                 Console.Write("testing a simple plug-in... ");
                 Console.Out.Flush();
-                await using Communicator communicator = Initialize(new Dictionary<string, string>
-                {
+                await using Communicator communicator = Initialize(
+                    new Dictionary<string, string>
                     {
-                        "Ice.Plugin.Test",
-                        $"{pluginPath }:ZeroC.Ice.Test.Plugin.PluginFactory 'C:\\Program Files\\' --DatabasePath 'C:\\Program Files\\Application\\db'"
-                    }
-                });
+                        {
+                            "Ice.Plugin.Test",
+                            $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginFactory 'C:\\Program Files\\' --DatabasePath 'C:\\Program Files\\Application\\db'"
+                        }
+                    });
                 await communicator.ActivateAsync();
                 Console.WriteLine("ok");
             }
@@ -38,12 +39,13 @@ namespace ZeroC.Ice.Test.Plugin
                 Console.Out.Flush();
                 try
                 {
-                    await using Communicator communicator = Initialize(new Dictionary<string, string>()
-                    {
+                    await using Communicator communicator = Initialize(
+                        new Dictionary<string, string>()
                         {
-                            "Ice.Plugin.Test", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginInitializeFailFactory"
-                        }
-                    });
+                            {
+                                "Ice.Plugin.Test", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginInitializeFailFactory"
+                            }
+                        });
                     await communicator.ActivateAsync();
                     Assert(false);
                 }
@@ -58,13 +60,14 @@ namespace ZeroC.Ice.Test.Plugin
                 Console.Write("testing plug-in load order... ");
                 Console.Out.Flush();
 
-                await using Communicator communicator = Initialize(new Dictionary<string, string>()
-                {
-                    { "Ice.Plugin.PluginOne", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginOneFactory" },
-                    { "Ice.Plugin.PluginTwo", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginTwoFactory" },
-                    { "Ice.Plugin.PluginThree", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginThreeFactory" },
-                    { "Ice.PluginLoadOrder", "PluginOne, PluginTwo" } // Exclude PluginThree
-                });
+                await using Communicator communicator = Initialize(
+                    new Dictionary<string, string>()
+                    {
+                        { "Ice.Plugin.PluginOne", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginOneFactory" },
+                        { "Ice.Plugin.PluginTwo", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginTwoFactory" },
+                        { "Ice.Plugin.PluginThree", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginThreeFactory" },
+                        { "Ice.PluginLoadOrder", "PluginOne, PluginTwo" } // Exclude PluginThree
+                    });
                 await communicator.ActivateAsync();
                 Console.WriteLine("ok");
             }
@@ -73,28 +76,19 @@ namespace ZeroC.Ice.Test.Plugin
                 Console.Write("testing plug-in manager... ");
                 Console.Out.Flush();
 
-                MyPlugin p4;
-                {
-                    await using Communicator communicator = Initialize(new Dictionary<string, string>()
+                await using Communicator communicator = Initialize(
+                    new Dictionary<string, string>()
                     {
                         { "Ice.Plugin.PluginOne", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginOneFactory" },
                         { "Ice.Plugin.PluginTwo", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginTwoFactory" },
                         { "Ice.Plugin.PluginThree", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginThreeFactory" },
                     });
 
-                    Assert(communicator.GetPlugin("PluginOne") != null);
-                    Assert(communicator.GetPlugin("PluginTwo") != null);
-                    Assert(communicator.GetPlugin("PluginThree") != null);
-
-                    p4 = new MyPlugin();
-                    communicator.AddPlugin("PluginFour", p4);
-                    Assert(communicator.GetPlugin("PluginFour") != null);
-
-                    await communicator.ActivateAsync();
-
-                    Assert(p4.IsInitialized());
-                }
-                Assert(p4.IsDestroyed());
+                Assert(communicator.Plugins.Count == 3);
+                Assert(communicator.Plugins[0].Name == "PluginOne");
+                Assert(communicator.Plugins[1].Name == "PluginTwo");
+                Assert(communicator.Plugins[2].Name == "PluginThree");
+                await communicator.ActivateAsync();
                 Console.WriteLine("ok");
             }
 
@@ -103,13 +97,14 @@ namespace ZeroC.Ice.Test.Plugin
                 Console.Out.Flush();
                 try
                 {
-                    await using Communicator communicator = Initialize(new Dictionary<string, string>()
-                    {
-                        { "Ice.Plugin.PluginOneFail", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginOneFailFactory" },
-                        { "Ice.Plugin.PluginTwoFail", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginTwoFailFactory" },
-                        { "Ice.Plugin.PluginThreeFail", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginThreeFailFactory" },
-                        { "Ice.PluginLoadOrder", "PluginOneFail, PluginTwoFail, PluginThreeFail"}
-                    });
+                    await using Communicator communicator = Initialize(
+                        new Dictionary<string, string>()
+                        {
+                            { "Ice.Plugin.PluginOneFail", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginOneFailFactory" },
+                            { "Ice.Plugin.PluginTwoFail", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginTwoFailFactory" },
+                            { "Ice.Plugin.PluginThreeFail", $"{pluginPath}:ZeroC.Ice.Test.Plugin.PluginThreeFailFactory" },
+                            { "Ice.PluginLoadOrder", "PluginOneFail, PluginTwoFail, PluginThreeFail"}
+                        });
                     await communicator.ActivateAsync();
                 }
                 catch
@@ -117,28 +112,6 @@ namespace ZeroC.Ice.Test.Plugin
                     // Expected
                 }
                 Console.WriteLine("ok");
-            }
-        }
-
-        internal class MyPlugin : IPlugin
-        {
-            private bool _destroyed;
-            private bool _initialized;
-
-            public bool IsInitialized() => _initialized;
-
-            public bool IsDestroyed() => _destroyed;
-
-            public Task ActivateAsync(CancellationToken cancel)
-            {
-                _initialized = true;
-                return Task.CompletedTask;
-            }
-
-            public ValueTask DisposeAsync()
-            {
-                _destroyed = true;
-                return new ValueTask(Task.CompletedTask);
             }
         }
     }
